refactor: storage layer to use postgres dockertest (#114)
* feat(store): add dockertest postgres

* refactor(provider): use postgres to test and pass context

* refactor(namespace): use postgres to test and pass context

* refactor(receiver): use postgres to test and pass context

* refactor(alert): use postgres to test and pass context

* refactor(template): use postgres to test and pass context

* refactor(rule list): use postgres to test and pass context

* refactor: change all slice of pointer struct to slice of struct

* refactor(rule): simplify transaction usage

* refactor(subscription): use postgres to test and simplify transaction

* refactor(postgres): decouple postgres from implementation by using abstraction

* refactor: improve error and log

* refactor: create and update return error only

* refactor: revert transaction approach

+++ cli/template.go
@@ -222,7 +222,7 @@ func getTemplateCmd(c *configuration) *cobra.Command {
 			}
 
 			template := &template.Template{
-				ID:        uint(templateData.GetId()),
+				ID:        templateData.GetId(),
 				Name:      templateData.GetName(),
 				Body:      templateData.GetBody(),
 				Tags:      templateData.GetTags(),
